+++ device-simulator/device_simulator/models.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class ClimateState(BaseModel):
     power: bool | None = None
-    target_temperature: float | None = None
+    target_temperature: int | None = None
 
 
 class Command(BaseModel):
